refactor: componentize LearnArticles from features page
